Add unit test for runtime entrypoint dependencies in test_dependencies.py
- Created a new test file to verify the import of essential dependencies: requests and tqdm.
- Added assertions to check the versions of the imported libraries and ensure the download_nhanes_files function is callable.

+++ src/nhanes_feasibility/threshold_analysis.py
@@ -73,7 +73,7 @@ def testing_burden_for_capture_targets(
     rows: list[dict[str, object]] = []
 
     for target in targets:
-        required_cases = int(math.ceil(target * total_high_risk))
+        required_cases = math.ceil(target * total_high_risk)
         if required_cases == 0:
             tested_n = 0
         else:
